+++ src/main/java/load/datapool/service/BaseLocker.java
@@ -16,6 +16,7 @@ public class BaseLocker implements Locker {
     private int size = 0;
     private int bufferSize = 1000;
     public static final int startIndex = 1;
+    private boolean markedAsEmpty = false; // Mark true if the pool has become empty. There is no need to check table in the DB
 
     public BaseLocker(String poolName) {
         this.poolName = poolName;
@@ -31,12 +32,14 @@ public class BaseLocker implements Locker {
 
     @Override
     public void add() {
-        size++;
+        add(1);
+    }
+    public void add(int count) {
+        size += count;
         if (size >= list.length) {
             list = Arrays.copyOf(list, size+bufferSize);
         }
     }
-
     @Override
     public void lock(int id) {
         id-= startIndex;
@@ -69,11 +72,31 @@ public class BaseLocker implements Locker {
         logger.debug("Pool {}: id = {}, startIndex = {}, size = {}", getPoolName(), id, startIndex, size);
         id-= startIndex;
         if (id >= size) return 0;   // not exist
-        for (int i = id; i < list.length; i++) {
-            if (!list[i])
-                return i+startIndex;
+        for (int i = id; i < size; i++) {
+            if (!list[i]) {
+                logger.debug("Now:" + String.valueOf(list[i]));
+                logger.debug("Next: " + String.valueOf(list[i + 1]));
+                return i + startIndex;
+            }
         }
         return 0;
+    }
+
+    @Override
+    public boolean isMarkedAsEmpty() {
+        return this.markedAsEmpty;
+    }
+    @Override
+    public void markAsEmpty() {
+        if (this.markedAsEmpty == true) return;
+        logger.warn("{} marked like empty pool.", poolName);
+        this.markedAsEmpty = true;
+    }
+    @Override
+    public void markAsNotEmpty() {
+        if (this.markedAsEmpty == false) return;
+        logger.warn("{} marked like NOT empty pool.", poolName);
+        this.markedAsEmpty = false;
     }
 
 }